对第三方用户的登录注册测试。

+++ ATOMPSGod/ATOMPSGod/Controller/login/PIEVerificationCodeCountdownButton.m
@@ -86,7 +86,8 @@
              numberCount = numberLimit;
          }else{
              
-             NSString *countdownString = [NSString stringWithFormat:@"%ld秒后重发", numberCount];
+             NSString *countdownString =
+             [NSString stringWithFormat:@"%ld秒后重发", (long)numberCount];
              
              [self setTitle:countdownString
                                    forState:UIControlStateNormal];
